add #16 json取り扱い処理を別クラスへ切り分け
